splitting of binary. generate only needed for web server
