+++ cmd/stateexplorer/static/index.html
